Minor changes do validation logic
